1.补充cpr AO 域和PD 域的reg 说明。 2.修改kbs 模块中mtxkey的类型为uint16_t 。

+++ XC6xx_ble_sdk/modules/xincx/drivers/include/xincx_kbs.h
@@ -52,7 +52,7 @@ typedef void (*xincx_kbs_mtxkey_handler_t)(int16_t mtxkey_idx,uint16_t mtxkey_va
 /**@brief Mtxkey configuration structure. */
 typedef struct
 {
-    uint8_t              key_val;           /**< Pin set value. */
+    uint16_t             key_val;           /**< Pin set value. */
     uint8_t              row_pin;           /**< Pin to be used as a row button. */
     uint8_t              col_pin;           /**< Pin to be used as a col button. */
    // uint8_t              active_state;     /**< APP_BUTTON_ACTIVE_HIGH or APP_BUTTON_ACTIVE_LOW. */
